Tom has requested that the makecurrent checkbox not be automatically checked when a date is entered into the Date box in the "Add a New Determination" tab in the occurrence editor.  He thinks it is likely to lead to an erroneous replacement when somebody is working fast.  I'm inclined to agree so I eliminated the onchange="detDateChanged(this.form);" on line 185.
There is another alternative: The detDateChanged function which is defined in collections.occureditormain.js automatically checks the box when either the identification date is null or the entered date is later than the identified by date.  Eliminating the automatic check in the first case could reduce the number of potential errors.

It's up to you.