OccurrenceEditorManager.php: Modified function editOccurrence to change exstitle to title in the editedFields array.  Changed the query on line 610 to change
'exstitle' to 'title'.

*************************************************
I have determined that these changes in OccurrenceEditorManager.php eliminate an error that was generated before when the NLP code found the exsiccati and 

that it doesn't interfere with autocomplete in the exsiccati title field.  I have seen no other adverse effects of making this change but we should watch for 

any adverse effects that I may not have noticed as this is not my code and I am not sure I understand all of the possible implications of making them.
*************************************************

SpecProcNlp.php: Added includes for SpecProcNlpParserLBCCLichen.php and SpecProcNlpParserLBCCBryophyte.php.  Changed the query in the parseRawOcrRecord 

function to retrieve the collid, url and catalognumber and added thes to the call to the parseTextBlock function.  I'm not sure if this method is ever used.  

Added code to the parseTextBlock function to assign the handler to be either SpecProcNlpParserLBCCBryophyte or SpecProcNlpParserLBCCLichen based on the URL 

or SpecProcNlpParserLBCCCommon if the organism type cannot be determined from the URL.  Passed the catalogNumber to the handler through the constructor.  

Also passed the collid to SpecProcNlpParserLBCCLichen through the constructor.  Called $handler->parse instead of $this->parse.  Similarly modified the 

batchProcess function to get the collid, url and catalogNumber and pass them to the parseTextBlock function.

SpecProcNlpParserLBCCCommon.php: Removed the includes for SpecProcNlpParserLBCCLichen.php and SpecProcNlpParserLBCCBryophyte.php as these are included in 

SpecProcNlp.php.  Added the class variable protected $conn;  Changed the constructor to set the catalogNumber and set up the connection.  Changed the 

destructor to close the connection.  Removed the code to set the handler as this is now done in SpecProcNlp.  Added code to get the exsiccati title and 

abbreviation.  Added private function containsVerbatimAttribute (from SpecProcNlpParserLBCCLichen.  This can changed later but I don't expect this to happen 

often) so that it is available when the generic method is used because the type of organism can't be deternmined from the URL.  Added private function 

getLabelInfo to call doGenericLabel when the type of organism can't be deternmined from the URL.  Modified function formatDate so that it adds "-00" to the 

date when has no day and "-00-00" when it has no month.  Modified function fixString so that it uses $catNo = $this->catalogNumber.  Changed $this-

>getLabelInfo to $handler->getLabelInfo.

SpecProcNlpParserLBCCLichen.php: Eliminated destructor so the parent will be used.  Modified the constructor to set the collid and pass the catalogNumber to 

the parent constructor.  Changed getLabelInfo and containsVerbatimAttribute from protected to private.  Modified getLabelInfo to use $this->collId.

SpecProcNlpParserLBCCBryophyte.php: Eliminated destructor so the parent will be used.  Modified the constructor to pass the catalogNumber to the parent 

constructor.  Changed getLabelInfo and containsVerbatimAttribute from protected to private.  Added doBryophytaNeotropicaExsiccataLabel, 

doBryophytaSelectaExsiccataLabel, doHepaticaeJaponicaeExsiccataeLabel and doBryophytorumTyporumExsiccataLabel along with isBryophytaNeotropicaExsiccataLabel, 

isBryophytaSelectaExsiccataLabel, isHepaticaeJaponicaeExsiccataeLabel and isBryophytorumTyporumExsiccataLabel to handle thes exsiccati.