Bug fix: Occurrence Upload Export
- Large uploads were failing to export (for review) due to memory
issues. The reason for this is because the full record set was first
being transferred into an array and then streamed out to browser. Code
was modified to simply stream out to browser directly from db recordset,
thus avoiding full dataset unnecessarily being stored in an array
variable prior to output.