Modification to CSS versioning
- Moved main.css versioning variable into the symbini.php file. This
gives portal managers the ability to force a refresh of clients browsers
without needing to alter the symbbase.php file. Only portal managers
would need to force a refresh of main.css bcause they are the only ones
who can modify this template file.
- Left base.css versioning control within the symbbase.php file.
Programmers are the central folks that would modify the base.css file,
and thus the one needing control of forcing the refresh of this file