Collection Metadata adjustments
- Temporarily commented out “Publish to iDigBio” checkbox until we can
better figure out how to integrate it into the workflow. It's causing
confusion because folks think that iDigBio will automatically harvest
their data if box is checked, but it’s not actually doing anything.
iDigBio’s current workflow requires that the collection needs to first
contact iDigBio directly requesting to be added into the queue. If we
want iDigBio to make use of the value in this field, we’ll need to add
it to the RSS feed and ask them to integrate this into their workflow.
GBIF checkbox is still activated.