Add support for on-the-fly image resizing for web & thumbnail images
This adds a new file imagelib/imgresize.php which can be passed a location for a JPEG image, a width, and an image quality.
The script will serve up that resized image on-demand using ImageMagic or GD.

In addition to the new file, this required adding an additional menu option for web and thumbnail images in collections/specprocessor/imageprocessor.php, and an additional field createWebImg in the database table specprocessorprojects.